requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.40.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
model-mommy==2.0.0
oauthlib==3.1.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.1.1
urllib3==1.25.7
zope.interface==5.0.1
